refactor: move library checks to GenericSpectrumLoader
theoretically a library would be needed for extraction in any
SpectrumLoader, (not just .mzml) so move the library checks to the
abstract class
