+++ massdash/loaders/MzMLDataLoader.py
@@ -3,14 +3,11 @@ massdash/loaders/MzMLDataLoader
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 """
 
-from os.path import basename, splitext
-from typing import Dict, List, Union, Literal
+from typing import Dict, List, Union
 import numpy as np
 import pandas as pd
 
 # Loaders
-from .access import MzMLDataAccess, OSWDataAccess
-from .SpectralLibraryLoader import SpectralLibraryLoader
 from .GenericSpectrumLoader import GenericSpectrumLoader
 from .ResultsLoader import ResultsLoader
 # Structs
@@ -29,25 +26,11 @@ class MzMLDataLoader(GenericSpectrumLoader):
         libraryFile: (str) The path to the library file (.tsv or .pqp)
         
     '''
-    def __init__(self, libraryFile, **kwargs):
+    def __init__(self, **kwargs):
         super().__init__(**kwargs) 
         self.dataAccess = [MzMLDataAccess(f, 'ondisk', verbose=self.verbose) for f in self.dataFiles]
         self.has_im = np.all([d.has_im for d in self.dataAccess])
-        self.libraryFile = libraryFile
-        self.libraryAccess = None
 
-        # If the library is not explicitly set, 
-        if self.libraryFile is not None:
-            self.libraryAccess = SpectralLibraryLoader(self.libraryFile)
-        else: # self.libraryFile is None:
-            for a in self.rsltsAccess:
-                if isinstance(a, OSWDataAccess): 
-                   self.libraryAccess = SpectralLibraryLoader(a.filename)
-        
-        # If library access is not set, then throw an error
-        if self.libraryAccess is None:
-            raise ValueError("If .osw file is not supplied, library file is required for MzMLDataLoader to perform targeted extraction")
-                   
     @ResultsLoader.cache_results
     def loadTransitionGroups(self, pep_id: str, charge: int, config: TargetedDIAConfig, runNames: Union[None, str, List[str]]=None) -> Dict[str, TransitionGroup]:
         '''
